+++ src/briefcase/config.py
@@ -869,7 +869,7 @@ def _normalize_pep639_license_config(
     if raw_license_files is None:
         raw_license_files = []
 
-    # Ensure `licence` is an SPDX expression
+    # Ensure `license` is an SPDX expression
     try:
         spdx_id = canonicalize_license_expression(raw_license)
     except InvalidLicenseExpression:
@@ -917,65 +917,61 @@ def _normalize_pep621_license_text_config(
     # Attempt to identify the SPDX expression from the text content.
     spdx_id = get_license_from_text(license_text)
 
-    warning = [
-        f"""
-*******************************************************************************
-** {"WARNING: '" + app_name + "' uses PEP 621 `license.text` format":73} **
-*******************************************************************************
-
-    Briefcase now uses PEP 639 format for license definitions.
-"""
-    ]
     if spdx_id is not None:
         # SPDX identifiable.
         license = spdx_id
-        warning.append(f"""
-    PEP 639 requires the definition of both `license` and `license-files`,
-    and `license` must be a valid SPDX expression. The current value for
-    `license.text` seems to define a SPDX license of '{spdx_id}'.
-""")
+        spdx_note = (
+            "PEP 639 requires the definition of both `license` and "
+            "`license-files`, and `license` must be a valid SPDX expression. "
+            "The current value for `license.text` seems to define a SPDX "
+            f"license of '{spdx_id}'."
+        )
     else:
         # SPDX not identifiable.
         spdx_id = "<SPDX expression>"
         license = "LicenseRef-UnknownLicense"
-        warning.append(f"""
-    PEP 639 requires the definition of both `license` and `license-files`.
-    Briefcase cannot determine the current license for '{app_name}' based
-    on the value of `license.text`. A value of 'LicenseRef-UnknownLicense'
-    will be used.
-""")
+        spdx_note = (
+            "PEP 639 requires the definition of both `license` and "
+            "`license-files`. Briefcase cannot determine the current license "
+            f"for '{app_name}' based on the value of `license.text`. A value of "
+            "'LicenseRef-UnknownLicense' will be used."
+        )
 
     # Write the license text to a file under the build directory so it can
     # be referenced as a real path in license-files.
     tmp_license_file = _write_temp_license(base_path, app_name, license_text)
     if tmp_license_file:
         license_files = [tmp_license_file]
-        warning.append("""
-    The contents of `license.text` will be used as the contents of the
-    license file. This may not be correct, and should be verified.
-""")
+        file_note = (
+            "The contents of `license.text` will be used as the contents of "
+            "the license file. This may not be correct, and should be verified."
+        )
     else:
         license_files = []
-        warning.append("""
-    Your project will not have a value for `license-files`. This will
-    cause problems packaging for some platforms.
-""")
+        file_note = (
+            "Your project will not have a value for `license-files`. This will "
+            "cause problems packaging for some platforms."
+        )
 
-    warning.append(f"""
-    Update your configuration to put the full license text in a file and use
-    PEP 639 format for the license definition:
+    # Warn and finalize PEP 621 license.text coercion.
+    console.warning_banner(
+        f"'{app_name}' uses PEP 621 `license.text` format",
+        f"""
+            Briefcase now uses PEP 639 format for license definitions.
 
-        license = "{spdx_id}"
-        license-files = ["LICENSE"]
+            {spdx_note}
 
-    You should not release your project without resolving this warning.
+            {file_note}
 
-*******************************************************************************
-""")
+            Update your configuration to put the full license text in a file
+            and use PEP 639 format for the license definition:
 
-    # Warn and finalize PEP 621 license.text coercion. Use `license_files` rather than
-    # `license-files` so it's a valid attribute name.
-    console.warning("".join(warning))
+                license = "{spdx_id}"
+                license-files = ["LICENSE"]
+
+            You should not release your project without resolving this warning.
+        """,
+    )
     config["license"] = license
     config["license_files"] = license_files
 
@@ -1006,47 +1002,41 @@ def _normalize_pep621_license_file_config(
     license_text = license_path.read_text(encoding="utf-8")
     spdx_id = get_license_from_text(license_text)
 
-    warning = [
-        f"""
-*******************************************************************************
-** {"WARNING: '" + app_name + "' uses PEP 621 `license.file` format":73} **
-*******************************************************************************
-
-    Briefcase now uses PEP 639 format for license definitions.
-
-    PEP 639 requires the definition of both `license` and `license-files`.
-    The value for `license.file` will be used to populate the PEP 639
-    `licence-files` setting.
-"""
-    ]
     if spdx_id is not None:
         license = spdx_id
         # License is valid SPDX
-        warning.append(f"""
-    The license has been identified as '{spdx_id}'.
-""")
+        spdx_note = f"The license has been identified as '{spdx_id}'."
     else:
         # Can't identify SPDX for license
         license = "<SPDX expression>"
         spdx_id = "LicenseRef-UnknownLicense"
-        warning.append("""
-    A license SPDX expression could not be identified from the license file.
-    The license has been set to 'LicenseRef-UnknownLicense'
-    """)
+        spdx_note = (
+            "A license SPDX expression could not be identified from the "
+            "license file. The license has been set to "
+            "'LicenseRef-UnknownLicense'."
+        )
 
-    warning.append(f"""
-    Update your configuration to use PEP 639 format:
-
-        license = "{license}"
-        license-files = ["{license_file}"]
-
-    You should not release your project without resolving this warning.
-
-*******************************************************************************
-""")
     # Warn and finalize PEP 621 license.file coercion. Use `license_files` rather than
     # `license-files` so it's a valid attribute name.
-    console.warning("".join(warning))
+    console.warning_banner(
+        f"'{app_name}' uses PEP 621 `license.file` format",
+        f"""
+            Briefcase now uses PEP 639 format for license definitions.
+
+            PEP 639 requires the definition of both `license` and
+            `license-files`. The value for `license.file` will be used to
+            populate the PEP 639 `license-files` setting.
+
+            {spdx_note}
+
+            Update your configuration to use PEP 639 format:
+
+                license = "{license}"
+                license-files = ["{license_file}"]
+
+            You should not release your project without resolving this warning.
+        """,
+    )
     config["license"] = spdx_id
     config["license_files"] = [license_file]
 
@@ -1076,65 +1066,62 @@ def _normalize_pre_pep621_license_config(
     # Attempt to identify the SPDX expression from the text content.
     spdx_id = get_license_from_text(license_text)
 
-    warning = [
-        f"""
-*******************************************************************************
-** {"WARNING: '" + app_name + "' uses pre-PEP 621 `license` format":73} **
-*******************************************************************************
-
-    Briefcase now uses PEP 639 format for license definitions.
-"""
-    ]
     if spdx_id is not None:
         # SPDX identifiable.
         license = spdx_id
-        warning.append(f"""
-    PEP 639 requires the definition of both `license` and `license-files`,
-    and `license` must be a valid SPDX expression. The current value for
-    `license` seems to define a SPDX license of '{spdx_id}'.
-""")
+        spdx_note = (
+            "PEP 639 requires the definition of both `license` and "
+            "`license-files`, and `license` must be a valid SPDX expression. "
+            "The current value for `license` seems to define a SPDX license "
+            f"of '{spdx_id}'."
+        )
     else:
         # SPDX not identifiable.
         spdx_id = "<SPDX expression>"
         license = "LicenseRef-UnknownLicense"
-        warning.append(f"""
-    PEP 639 requires the definition of both `license` and `license-files`.
-    Briefcase cannot determine the current license for '{app_name}' based
-    on the value of `license`. A value of 'LicenseRef-UnknownLicense' will
-    be used.
-""")
+        spdx_note = (
+            "PEP 639 requires the definition of both `license` and "
+            "`license-files`. Briefcase cannot determine the current license "
+            f"for '{app_name}' based on the value of `license`. A value of "
+            "'LicenseRef-UnknownLicense' will be used."
+        )
 
     # Write the license text to a file under the build directory so it can
     # be referenced as a real path in license-files.
     tmp_license_file = _write_temp_license(base_path, app_name, license_text)
     if tmp_license_file:
         license_files = [tmp_license_file]
-        warning.append("""
-    The contents of `license` will be used as the contents of the license
-    file. This may not be correct, and should be verified.
-""")
+        file_note = (
+            "The contents of `license` will be used as the contents of the "
+            "license file. This may not be correct, and should be verified."
+        )
     else:
         license_files = []
-        warning.append("""
-    Your project will not have a value for `license-files`. This will
-    cause problems packaging for some platforms.
-""")
-
-    warning.append(f"""
-    Update your configuration to put the full license text in a file and use
-    PEP 639 format for the license definition:
-
-        license = "{spdx_id}"
-        license-files = ["LICENSE"]
-
-    You should not release your project without resolving this warning.
-
-*******************************************************************************
-""")
+        file_note = (
+            "Your project will not have a value for `license-files`. This will "
+            "cause problems packaging for some platforms."
+        )
 
     # Warn and finalize pre-PEP 621 license coercion. Use `license_files` rather than
     # `license-files` so it's a valid attribute name.
-    console.warning("".join(warning))
+    console.warning_banner(
+        f"'{app_name}' uses pre-PEP 621 `license` format",
+        f"""
+            Briefcase now uses PEP 639 format for license definitions.
+
+            {spdx_note}
+
+            {file_note}
+
+            Update your configuration to put the full license text in a file
+            and use PEP 639 format for the license definition:
+
+                license = "{spdx_id}"
+                license-files = ["LICENSE"]
+
+            You should not release your project without resolving this warning.
+        """,
+    )
     config["license"] = license
     config["license_files"] = license_files
 
@@ -1476,6 +1463,21 @@ def parse_config(config_file: Path, platform, output_format, console):
 
         # Normalize license fields to PEP 639 representation.
         normalize_license_config(config, app_name, base_path, console)
+
+        # Warn if the description is too long for some packaging formats.
+        description = config.get("description")
+        if isinstance(description, str) and len(description) > 80:
+            console.warning_banner(
+                "Application description is too long",
+                f"""
+                    The description for {app_name!r} is {len(description)}
+                    characters long. Briefcase recommends a description of no
+                    more than 80 characters; longer descriptions may be
+                    truncated when packaging for some platforms.
+
+                    Move any detailed text into the `long_description` field.
+                """,
+            )
 
         # Construct a configuration object, and add it to the list
         # of configurations that are being handled.