+++ src/briefcase/platforms/linux/flatpak.py
@@ -97,6 +97,21 @@ Your application configuration must provide values for
 `flatpak_sdk`, `flatpak_runtime`, and `flatpak_runtime_version`.
 """) from e
 
+    def flatpak_base(self, app):
+        return getattr(app, "flatpak_base", None)
+
+    def flatpak_base_version(self, app):
+        try:
+            return app.flatpak_base_version
+        except AttributeError as e:
+            if self.flatpak_base(app) is not None:
+                raise BriefcaseConfigError("""\
+The App specifies a Flatpak base without a version.
+
+Your application configuration must provide a value for `flatpak_base_version`
+if `flatpak_base` is defined.
+""") from e
+
 
 class LinuxFlatpakCreateCommand(LinuxFlatpakMixin, CreateCommand):
     description = "Create and populate a Linux Flatpak."
@@ -108,6 +123,8 @@ class LinuxFlatpakCreateCommand(LinuxFlatpakMixin, CreateCommand):
             "flatpak_runtime": self.flatpak_runtime(app),
             "flatpak_runtime_version": self.flatpak_runtime_version(app),
             "flatpak_sdk": self.flatpak_sdk(app),
+            "flatpak_base": self.flatpak_base(app),
+            "flatpak_base_version": self.flatpak_base_version(app),
         }
 
     def permissions_context(
@@ -185,6 +202,8 @@ class LinuxFlatpakBuildCommand(LinuxFlatpakMixin, BuildCommand):
             runtime=self.flatpak_runtime(app),
             runtime_version=self.flatpak_runtime_version(app),
             sdk=self.flatpak_sdk(app),
+            base=self.flatpak_base(app),
+            base_version=self.flatpak_base_version(app),
         )
 
         self.console.info("Building Flatpak...", prefix=app.app_name)